requirements/base.txt
# Django
Django==3.2
djangorestframework==3.12.4
django-ckeditor==6.1.0
django-solo==1.1.5

# Files
whitenoise==5.3.0
Pillow==8.3.1

# postgres
psycopg2-binary==2.9.1
